+++ src/uniforms/LightsInfoUniformStruct.js
@@ -4,6 +4,8 @@ const LIGHT_PIXELS = 6;
 const RECT_AREA_LIGHT = 0;
 const CIRC_AREA_LIGHT = 1;
 const SPOT_LIGHT = 2;
+const DIR_LIGHT = 3;
+const POINT_LIGHT = 4;
 export class LightsInfoUniformStruct {
 
 	constructor() {
@@ -69,6 +71,14 @@ export class LightsInfoUniformStruct {
 			} else if ( l.isSpotLight ) {
 
 				type = SPOT_LIGHT;
+
+			} else if ( l.isDirectionalLight ) {
+
+				type = DIR_LIGHT;
+
+			} else if ( l.isPointLight ) {
+
+				type = POINT_LIGHT;
 
 			}
 
@@ -157,6 +167,33 @@ export class LightsInfoUniformStruct {
 
 				// iesProfile
 				floatArray[ baseIndex + ( index ++ ) ] = iesTextures.indexOf( l.iesTexture );
+
+			} else if ( l.isPointLight ) {
+
+				const worldPosition = l.getWorldPosition( u );
+				floatArray[ baseIndex + ( index ++ ) ] = worldPosition.x;
+				floatArray[ baseIndex + ( index ++ ) ] = worldPosition.y;
+				floatArray[ baseIndex + ( index ++ ) ] = worldPosition.z;
+				index ++;
+
+				// sample 4
+				index += 4;
+
+				// sample 5
+				index += 2;
+
+				floatArray[ baseIndex + ( index ++ ) ] = l.decay;
+				floatArray[ baseIndex + ( index ++ ) ] = l.distance;
+
+			} else if ( l.isDirectionalLight ) {
+
+				const worldPosition = l.getWorldPosition( u );
+				const targetPosition = l.target.getWorldPosition( v );
+
+				target.subVectors( worldPosition, targetPosition ).normalize();
+				floatArray[ baseIndex + ( index ++ ) ] = target.x;
+				floatArray[ baseIndex + ( index ++ ) ] = target.y;
+				floatArray[ baseIndex + ( index ++ ) ] = target.z;
 
 			}