+++ CHANGELOG.md
@@ -5,9 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Equirect sampling CDF offset values causing env maps with 1 bright pixel to be most noticeably incorrect.
+- Clearcoat roughness map values not being respected.
+
+## [0.0.9] - 2022-12-31
 ### Added
 - Support for `Material.flatShading` to render flat-shaded materials.
 - Support for randomization using Owen-scrambled and shuffled Sobol values enabling sample stratification and image in fewer samples.
+- Support for directional lights, point lights.
 
 ### Fixed
 - Roughness and metalness maps not being assigned correctly.